Move LoadReportingServiceServerBuilderOption from ::grpc_impl to ::grpc
Reverts: https://github.com/grpc/grpc/pull/18419/files
