+++ include/grpcpp/ext/server_load_reporting.h
@@ -19,22 +19,33 @@
 #ifndef GRPCPP_EXT_SERVER_LOAD_REPORTING_H
 #define GRPCPP_EXT_SERVER_LOAD_REPORTING_H
 
-#include <grpcpp/ext/server_load_reporting_impl.h>
+#include <grpc/support/port_platform.h>
+
+#include <grpc/load_reporting.h>
+#include <grpcpp/impl/codegen/config.h>
+#include <grpcpp/impl/codegen/server_context.h>
+#include <grpcpp/impl/server_builder_option.h>
 
 namespace grpc {
 namespace load_reporter {
 namespace experimental {
 
-typedef ::grpc_impl::load_reporter::experimental::
-    LoadReportingServiceServerBuilderOption
-        LoadReportingServiceServerBuilderOption;
+// The ServerBuilderOption to enable server-side load reporting feature. To
+// enable the feature, please make sure the binary builds with the
+// grpcpp_server_load_reporting library and set this option in the
+// ServerBuilder.
+class LoadReportingServiceServerBuilderOption
+    : public grpc::ServerBuilderOption {
+ public:
+  void UpdateArguments(::grpc::ChannelArguments* args) override;
+  void UpdatePlugins(std::vector<std::unique_ptr<::grpc::ServerBuilderPlugin>>*
+                         plugins) override;
+};
 
-static inline void AddLoadReportingCost(grpc::ServerContext* ctx,
-                                        const std::string& cost_name,
-                                        double cost_value) {
-  ::grpc_impl::load_reporter::experimental::AddLoadReportingCost(ctx, cost_name,
-                                                                 cost_value);
-}
+// Adds the load reporting cost with \a cost_name and \a cost_value in the
+// trailing metadata of the server context.
+void AddLoadReportingCost(grpc::ServerContext* ctx,
+                          const std::string& cost_name, double cost_value);
 
 }  // namespace experimental
 }  // namespace load_reporter
